# Parallel HashingVectorizer