--- a/benchmarks/bench_vectorizers.py
+++ b/benchmarks/bench_vectorizers.py
@@ -20,7 +20,8 @@ if __name__ == "__main__":
     print("# vectorizing {} documents:".format(len(data)))
 
     for label, vect in [
-        ("HashingVectorizer (vtext)", vtext.vectorize.HashingVectorizer()),
+        ("HashingVectorizer (vtext, n_jobs=1)", vtext.vectorize.HashingVectorizer()),
+        ("HashingVectorizer (vtext, n_jobs=2)", vtext.vectorize.HashingVectorizer(n_jobs=2)),
         (
             "HashingVectorizer (scikit-learn)",
             skt.HashingVectorizer(lowercase=False, norm=None),