--- a/python/vtext/vectorize.py
+++ b/python/vtext/vectorize.py
@@ -41,12 +41,13 @@ class HashingVectorizer(BaseEstimator):
         Whether the feature should be made of word or character n-grams.
 
         Currently only the "word" tokenizer is implemented.
-
     binary : boolean, default=False
         If True, all non zero counts are set to 1. This is useful for discrete
         probabilistic models that model binary events rather than integer
         counts.
-
+    n_jobs : int, default=1
+        number of threads to use for parallel feature extraction. Using n_jobs > 1,
+        is faster, but uses more memory.
     dtype : type, optional
         Type of the matrix returned by fit_transform() or transform().
 
@@ -67,10 +68,11 @@ class HashingVectorizer(BaseEstimator):
 
     """
 
-    def __init__(self, *, analyzer="word", binary=False, dtype=np.int32):
+    def __init__(self, *, analyzer="word", binary=False, n_jobs=1, dtype=np.int32):
         self.analyzer = analyzer
         self.binary = binary
         self.dtype = dtype
+        self.n_jobs = n_jobs
 
     def partial_fit(self, X, y=None):
         """Does nothing: this transformer is stateless.
@@ -90,6 +92,9 @@ class HashingVectorizer(BaseEstimator):
         if self.analyzer != "word":
             raise NotImplementedError
 
+        if not isinstance(self.n_jobs, int) or self.n_jobs < 1:
+            raise ValueError("n_jobs={} must be a integer >= 1".format(self.n_jobs))
+
     def fit(self, X, y=None):
         """Does nothing: this transformer is stateless.
 
@@ -104,6 +109,7 @@ class HashingVectorizer(BaseEstimator):
             )
 
         self._validate_params()
+        self._inner = _lib._HashingVectorizerWrapper(n_jobs=self.n_jobs)
 
         return self
 
@@ -129,7 +135,11 @@ class HashingVectorizer(BaseEstimator):
 
         self._validate_params()
 
-        indices, indptr, data = _lib._HashingVectorizerWrapper().transform(X)
+        if not hasattr(self, "_inner"):
+            # Initialize the wrapper
+            self.fit(None)
+
+        indices, indptr, data = self._inner.transform(X)
         if self.binary:
             data.fill(1)