--- a/src/lib.rs
+++ b/src/lib.rs
@@ -153,11 +153,17 @@ impl CountVectorizer {
 
         let tokenizer = tokenize::RegexpTokenizer::new(TOKEN_PATTERN_DEFAULT.to_string());
 
-        for (_document_id, document) in X.iter().enumerate() {
-            let document = document.to_ascii_lowercase();
+        // TODO: there should be a simpler way of handling this?
+        let tokenize = |doc: String| -> Vec<String> {
+            tokenizer.tokenize(&doc).map(|x| x.to_string()).collect()
+        };
 
-            let tokens = tokenizer.tokenize(&document);
+        let pipe = X
+            .iter()
+            .map(|doc| doc.to_ascii_lowercase())
+            .map(|doc| tokenize(doc));
 
+        for (_document_id, tokens) in pipe.enumerate() {
             indices_local.clear();
             for token in tokens {
                 let vocabulary_size = vocabulary.len() as i32;
@@ -228,15 +234,22 @@ impl HashingVectorizer {
 
         let tokenizer = tokenize::RegexpTokenizer::new(TOKEN_PATTERN_DEFAULT.to_string());
 
-        for (_document_id, document) in X.iter().enumerate() {
-            // String.to_lowercase() is very slow
-            // https://www.reddit.com/r/rust/comments/6wbru2/performance_issue_can_i_avoid_of_using_the_slow/
-            // https://github.com/rust-lang/rust/issues/26244
-            // Possibly use: https://github.com/JuliaStrings/utf8proc
-            // http://www.unicode.org/faq/casemap_charprop.html
-            let document = document.to_ascii_lowercase();
+        // TODO: there should be a simpler way of handling this?
+        let tokenize = |doc: String| -> Vec<String> {
+            tokenizer.tokenize(&doc).map(|x| x.to_string()).collect()
+        };
+        // String.to_lowercase() is very slow
+        // https://www.reddit.com/r/rust/comments/6wbru2/performance_issue_can_i_avoid_of_using_the_slow/
+        // https://github.com/rust-lang/rust/issues/26244
+        // Possibly use: https://github.com/JuliaStrings/utf8proc
+        // http://www.unicode.org/faq/casemap_charprop.html
+        //
+        let pipe = X
+            .iter()
+            .map(|doc| doc.to_ascii_lowercase())
+            .map(|doc| tokenize(doc));
 
-            let tokens = tokenizer.tokenize(&document);
+        for (_document_id, tokens) in pipe.enumerate() {
             indices_local.clear();
             for token in tokens {
                 let hash = fasthash::murmur3::hash32(&token);