--- a/src/vectorize/mod.rs
+++ b/src/vectorize/mod.rs
@@ -29,6 +29,7 @@ use crate::tokenize::Tokenizer;
 use hashbrown::HashMap;
 use ndarray::Array;
 use sprs::CsMat;
+use rayon::prelude::*;
 
 const TOKEN_PATTERN_DEFAULT: &str = r"\b\w\w+\b";
 
@@ -134,12 +135,21 @@ impl CountVectorizer {
 
         let tokenizer = tokenize::RegexpTokenizer::new(TOKEN_PATTERN_DEFAULT.to_string());
 
-        let pipe = X.iter().map(|doc| doc.to_ascii_lowercase());
 
         let mut vocabulary_size: i32 = 0;
 
-        for (_document_id, document) in pipe.enumerate() {
-            let tokens = tokenizer.tokenize(&document);
+        let tokenize = |doc: String| -> Vec<String> {
+            tokenizer.tokenize(&doc).map(|x| x.to_string()).collect()
+        };
+
+        let pipe: Vec<Vec<String>> = X
+            .par_iter()
+            .map(|doc| doc.to_ascii_lowercase())
+            .map(|doc| tokenize(doc))
+            .collect();
+
+
+        for tokens in pipe.iter() {
 
             indices_local.clear();
             for token in tokens {
@@ -212,10 +222,19 @@ impl HashingVectorizer {
         // https://github.com/rust-lang/rust/issues/26244
         // Possibly use: https://github.com/JuliaStrings/utf8proc
         // http://www.unicode.org/faq/casemap_charprop.html
-        let pipe = X.iter().map(|doc| doc.to_ascii_lowercase());
+        
+        let tokenize = |doc: String| -> Vec<String> {
+            tokenizer.tokenize(&doc).map(|x| x.to_string()).collect()
+        };
 
-        for (_document_id, document) in pipe.enumerate() {
-            let tokens = tokenizer.tokenize(&document);
+
+        let pipe: Vec<Vec<String>> = X
+            .par_iter()
+            .map(|doc| doc.to_ascii_lowercase())
+            .map(|doc| tokenize(doc))
+            .collect();
+
+        for tokens in pipe.iter() {
             indices_local.clear();
             for token in tokens {
                 // set the RNG seeds to get reproducible hashing