@rth 2019-05-08
In the end, this PR contains only the parallel version of the HashingVectorizer. CountVectorizer could be parallelized in a follow-up PR, the situation there is more complicated as it is not stateless and the vocabulary needs to be passed to different threads.

For HashingVectorizer, the scaling is reasonably good up to 8-16 CPU cores, after that we reach seem to reach the strong scaling limit, at least for this dataset. The maximum speed-up obtain is x5 of the scalar version. For `n_jobs=1` we fall back to the non parallelized version,

```
# vectorizing 19924 documents:
     HashingVectorizer (scikit-learn): 4.75s [19.2 MB/s], shape=(19924, 1048576), nnz=4177915
     HashingVectorizer (vtext, n_jobs=1): 1.16s [78.2 MB/s], shape=(19924, 1048576), nnz=3961670
     HashingVectorizer (vtext, n_jobs=2): 0.66s [137.4 MB/s], shape=(19924, 1048576), nnz=3961670
     HashingVectorizer (vtext, n_jobs=4): 0.40s [227.0 MB/s], shape=(19924, 1048576), nnz=3961670
     HashingVectorizer (vtext, n_jobs=8): 0.27s [340.0 MB/s], shape=(19924, 1048576), nnz=3961670
     HashingVectorizer (vtext, n_jobs=16): 0.23s [400.7 MB/s], shape=(19924, 1048576), nnz=3961670
     HashingVectorizer (vtext, n_jobs=32): 0.23s [394.6 MB/s], shape=(19924, 1048576), nnz=3961670
```
tested on EC2 c4.8xlarge with 36 CPU cores, and loading files from a tmpfs to avoid disk IO limiations.


The only limitation is that currently `n_jobs>1` uses all available CPU cores instead of the desired number. It could be adjusted at start time with `RAYON_NUM_THREADS` env variable. Fixing this would require using a local rayon thread pool instead of the global one. I have not found out how to do that with the current pipeline definition yet.